refactor(starterkit): split the number converters into one type per file
- MathConverters, RangeConverters and UnityNumberConverters become one file
  per type, matching how the rest of the StarterKit converters are laid out.
- Enum member summaries move to the multiline form the codebase uses
  everywhere else.
- Fully qualify the AudioMixer cref so it resolves without a using that
  nothing else in the file needs.
