+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Numbers/ClampMode.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 497376b6f54e4f25b783f549218b5ee4
+timeCreated: 1786988017
