+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Numbers/ClampNumberConverter.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 1de5756129154085934bfac477ba6f77
+timeCreated: 1786988043
