+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Numbers/CountdownProgressConverter.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: e11e30af2e8a464893875ca2f1d2fae5
+timeCreated: 1786988552
