+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Numbers/InverseLerpConverter.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: bee875b98f5048c4a33aead1c102dc8a
+timeCreated: 1786988510
