+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Numbers/LerpNumberConverter.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 3b312cbfd3d845009fbab7960f403d37
+timeCreated: 1786988519
