+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Numbers/NormalizedToPercentConverter.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 9d5da3ecd3eb4ab1a04a4755137f369d
+timeCreated: 1786988529
