+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Numbers/RoundMode.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: a58a08650f804b28a632d59befc4a660
+timeCreated: 1786988107
