+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Numbers/RoundNumberConverter.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: ff74bd4aa03c4257b7f93b0365137d4e
+timeCreated: 1786988118
