+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Numbers/SnapToStepConverter.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 887c4dcf317243efa7bdb730cb668e3a
+timeCreated: 1786988153
