+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Numbers/UnaryMathOperation.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 7376c953404a49ac960911d6197b0d6e
+timeCreated: 1786988171
