+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Numbers/WrapMode.cs
@@ -1,0 +1,19 @@
+// ReSharper disable once CheckNamespace
+namespace Aspid.MVVM.StarterKit
+{
+    /// <summary>
+    /// How <see cref="WrapNumberConverter"/> folds a value back into its range.
+    /// </summary>
+    public enum WrapMode
+    {
+        /// <summary>
+        /// Past the end, start again from the beginning.
+        /// </summary>
+        Repeat,
+
+        /// <summary>
+        /// Past the end, travel back towards the beginning.
+        /// </summary>
+        PingPong,
+    }
+}
