+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Unity/Runtime/Converters/Numbers/AnimationCurveConverter.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 72d8bc8576d34c5abd08673273f78796
+timeCreated: 1786987760
