+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Unity/Runtime/Converters/Numbers/AudioLinearToDecibelConverter.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 1c6a0cab6feb4eaba629caf831b98627
+timeCreated: 1786987833
